# docs/api/swagger.yaml
basePath: /v1
definitions:
  circle.CircleCurrency:
    enum:
    - USDC
    - XLM
    type: string
    x-enum-varnames:
    - CurrencyUSDC
    - CurrencyXLM
  circle.CircleFrequency:
    enum:
    - daily
    - weekly
    - biweekly
    - monthly
    type: string
    x-enum-varnames:
    - FrequencyDaily
    - FrequencyWeekly
    - FrequencyBiweekly
    - FrequencyMonthly
  circle.CircleType:
    enum:
    - public
    - private
    type: string
    x-enum-varnames:
    - CircleTypePublic
    - CircleTypePrivate
  circle.CreateCircleInput:
    properties:
      circleType:
        allOf:
        - $ref: '#/definitions/circle.CircleType'
        enum:
        - public
        - private
      collateralPercent:
        maximum: 100
        minimum: 0
        type: number
      contributionAmount:
        type: number
      currency:
        allOf:
        - $ref: '#/definitions/circle.CircleCurrency'
        enum:
        - USDC
        - XLM
      description:
        type: string
      frequency:
        allOf:
        - $ref: '#/definitions/circle.CircleFrequency'
        enum:
        - daily
        - weekly
        - biweekly
        - monthly
      gracePeriodHours:
        minimum: 0
        type: integer
      lateFeePercent:
        maximum: 100
        minimum: 0
        type: number
      maxMembers:
        maximum: 100
        minimum: 2
        type: integer
      maxStrikes:
        maximum: 10
        minimum: 1
        type: integer
      minMoiScore:
        maximum: 1000
        minimum: 0
        type: integer
      name:
        maxLength: 100
        minLength: 3
        type: string
      payoutType:
        allOf:
        - $ref: '#/definitions/circle.PayoutType'
        enum:
        - random
        - fixed
        - auction
        - vote
    required:
    - circleType
    - contributionAmount
    - currency
    - frequency
    - maxMembers
    - name
    - payoutType
    type: object
  circle.PayoutType:
    enum:
    - random
    - fixed
    - auction
    - vote
    type: string
    x-enum-varnames:
    - PayoutTypeRandom
    - PayoutTypeFixed
    - PayoutTypeAuction
    - PayoutTypeVote
  invite.GenerateInput:
    properties:
      circleId:
        type: string
      maxUses:
        maximum: 100
        minimum: 1
        type: integer
      ttlHours:
        maximum: 720
        minimum: 1
        type: integer
      userId:
        type: string
    required:
    - circleId
    - userId
    type: object
  response.Envelope:
    properties:
      data: {}
      error:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      success:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
host: moistello.com
info:
  contact:
    email: support@moistello.com
    name: Moistello Support
    url: https://moistello.com/support
  description: Decentralized savings circles on Stellar. REST API for circles, contributions,
    payouts, reputation, and governance.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://moistello.com/terms
  title: Moistello API
  version: 1.0.0
paths:
...
  /circles:
    get:
      description: Returns a paginated list of savings circles with optional search,
        status, and type filters.
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Filter by status
        enum:
        - pending
        - active
        - completed
        - cancelled
        in: query
        name: status
        type: string
      - description: Filter by type
        enum:
        - fixed
        - flexible
        - auction
        in: query
        name: type
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    circles:
                      type: array
                  type: object
                meta:
                  $ref: '#/definitions/response.PaginationMeta'
              type: object
      summary: List circles
      tags:
      - Circles
    post:
      consumes:
      - application/json
      description: Creates a new savings circle. Requires authentication.
      parameters:
      - description: Circle configuration
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/circle.CreateCircleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    circle:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Create a circle
      tags:
      - Circles
  /circles/{id}:
    get:
      description: Returns a single savings circle by ID.
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    circle:
                      type: object
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Get a circle
      tags:
      - Circles
  /circles/{id}/invites:
    get:
      description: Lists all invite codes for a circle. Circle owner only.
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    invites:
                      type: array
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: List circle invites
      tags:
      - Invites
    post:
      consumes:
      - application/json
      description: Generates a new invite code for a circle. Circle owner only.
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite settings (maxUses, expiresInDays)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/invite.GenerateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    invite:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Create an invite
      tags:
      - Invites
  /circles/{id}/join:
    post:
      consumes:
      - application/json
      description: Joins an existing savings circle. Requires an invite code if the
        circle is private.
      parameters:
      - description: Circle ID
        in: path
        name: id
        required: true
        type: string
      - description: Invite code (optional for public circles)
        in: body
        name: body
        schema:
          properties:
            inviteCode:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    success:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Join a circle
      tags:
      - Circles
  /contributions:
    get:
      description: Returns paginated contributions filtered by userId, circleId, or
        the authenticated user's own contributions.
      parameters:
      - description: Filter by user ID
        in: query
        name: userId
        type: string
      - description: Filter by circle ID
        in: query
        name: circleId
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    contributions:
                      type: array
                  type: object
                meta:
                  $ref: '#/definitions/response.PaginationMeta'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: List contributions
      tags:
      - Contributions
  /contributions/{id}:
    get:
      description: Returns a single contribution by ID.
      parameters:
      - description: Contribution ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    contribution:
                      type: object
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Get a contribution
      tags:
      - Contributions
...
  /invites/{code}:
    delete:
      description: Revokes an invite code so it can no longer be used.
      parameters:
      - description: Invite code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Envelope'
            - properties:
                data:
                  properties:
                    success:
                      type: boolean
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Revoke an invite
      tags:
      - Invites
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT token obtained from /auth/verify or /auth/register. Format:
      "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"